Changed all fluid API amounts and capacities to long.

+++ src/main/java/net/zarathul/simplemodslib/api/fluid/FluidContainerComponent.java
@@ -6,22 +6,22 @@ import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 
-public record FluidContainerComponent(FluidStack fluid, int capacity, boolean singleBucketMode)
+public record FluidContainerComponent(FluidStack fluid, long capacity, boolean singleBucketMode)
 {
-	public FluidContainerComponent(FluidStack fluid, int capacity)
+	public FluidContainerComponent(FluidStack fluid, long capacity)
 	{
 		this(fluid, capacity, true);
 	}
 
 	public static final Codec<FluidContainerComponent> CODEC = RecordCodecBuilder.create(instance -> instance.group(
 		FluidStack.CODEC.fieldOf("fluid").forGetter(FluidContainerComponent::fluid),
-		Codec.INT.fieldOf("capacity").forGetter(FluidContainerComponent::capacity),
+		Codec.LONG.fieldOf("capacity").forGetter(FluidContainerComponent::capacity),
 		Codec.BOOL.optionalFieldOf("singleBucketMode", true).forGetter(FluidContainerComponent::singleBucketMode)
 	).apply(instance, FluidContainerComponent::new));
 
 	public static final StreamCodec<RegistryFriendlyByteBuf, FluidContainerComponent> STREAM_CODEC = StreamCodec.composite(
 		FluidStack.STREAM_CODEC, FluidContainerComponent::fluid,
-		ByteBufCodecs.INT, FluidContainerComponent::capacity,
+		ByteBufCodecs.LONG, FluidContainerComponent::capacity,
 		ByteBufCodecs.BOOL, FluidContainerComponent::singleBucketMode,
 		FluidContainerComponent::new
 	);
